+++ ChangeLog.md
@@ -1,8 +1,12 @@
 ## 0.2.0.0
 
-* New `NodeStyle` type denoting flow/block style
-* `Event(SequenceStart,MappingStart)` constructors now record `NodeStyle`
-* `Style` type renamed to `ScalarType`
+* Changes in `Data.YAML.Event` module
+    * New `NodeStyle` type denoting flow/block style
+    * `Event(SequenceStart,MappingStart)` constructors now record `NodeStyle`
+    * `Style` type renamed to `ScalarType`
+    * New `writeEvents` and `writeEventsText` function
+
+---
 
 ### 0.1.1.2